Complimentary to previous commit.
I forgot few files.

+++ src/listmode_buildblock/LmToProjData.cxx
@@ -596,7 +596,7 @@ process_data()
 	   // When do_time_frame=true, the number of events is irrelevant, so we 
 	   // just set more_events to 1, and never change it
 	   long more_events = 
-	     do_time_frame? 1 : num_events_to_store;
+         do_time_frame? 1 : static_cast<long>(num_events_to_store);
 
 	   if (start_segment_index != proj_data_ptr->get_min_segment_num())
 	     {
